fix: lua close
Remove the deferred call to lua.LState.Close which was invalidating the
lua state before return.

In earlier versions of lua this wasn't an issue but in more recent versions
this causes a panic as it tries to use the nil stack field.

Update to the latest versions of all dependencies, which is particularly
important for lua which was 5 years old, missing many important bug
fixes.